Add vspipe option to print filter times

+++ src/vspipe/printgraph.h
@@ -25,6 +25,6 @@
 #include <string>
 
 std::string printFullNodeGraph(VSNodeRef *node, const VSAPI *vsapi);
-std::string printNodeTimes(VSNodeRef *node, int64_t processingTime, const VSAPI *vsapi);
+std::string printNodeTimes(VSNodeRef *node, double processingTime, const VSAPI *vsapi);
 
 #endif
